rename index file for indicators_table
