+++ x-pack/plugins/threat_intelligence/public/modules/indicators/components/indicators_table/indicators_table.tsx
